refactor(frontend): Move scenario codemirror mode to new file
