+++ frontend/src/modes/symbols.ts
@@ -1,0 +1,108 @@
+// keywords for fulibScenarios v1.7.0
+export const KEYWORDS = [
+  'a',
+  'add',
+  'adds',
+  'all',
+  'an',
+  'and',
+  'answer',
+  'answers',
+  'are',
+  'As', 'as',
+  'attribute',
+  'call',
+  'calls',
+  'card',
+  'cards',
+  // 'cf.',
+  'contain',
+  'contains',
+  'create',
+  'creates',
+  'do',
+  'does',
+  // 'e.g.',
+  'empty',
+  'equal',
+  'Every',
+  'expect',
+  'false',
+  'from',
+  'greater',
+  'has',
+  'have',
+  'in',
+  'into',
+  'is',
+  'it',
+  'less',
+  'like',
+  'link',
+  'many',
+  'match',
+  'matches',
+  'not',
+  'of',
+  'or',
+  'on',
+  'one',
+  'read',
+  'reads',
+  'register',
+  'remove',
+  'removes',
+  'same',
+  'some',
+  'take',
+  'takes',
+  'than',
+  'that',
+  'The', 'the',
+  'then',
+  'There', 'there',
+  'through',
+  'true',
+  'type',
+  'to',
+  'We', 'we',
+  'with',
+  'where',
+  'which',
+  'whose',
+  'write',
+  'writes',
+];
+
+export const TYPES = [
+  'void',
+  'boolean',
+  // numeric types
+  'byte',
+  'short',
+  'char',
+  'int',
+  'long',
+  'float',
+  'double',
+  // reference types
+  'Object',
+  'String',
+  'Number',
+  // wrapper types
+  'Void',
+  'Boolean',
+  'Byte',
+  'Short',
+  'Character',
+  'Integer',
+  'Long',
+  'Float',
+  'Double',
+  // lowercase variants
+  'object',
+  'string',
+  'number',
+  'character',
+  'integer',
+];
